% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphicalFunctions.R
\name{outlierHist}
\alias{outlierHist}
\title{a function for representing distribution using histogram and flag}
\usage{
outlierHist(datain, trait, flag, labelX)
}
\arguments{
\item{datain}{a dataframe}

\item{trait}{a variable to explore}

\item{flag}{a method of flag for outlier detection (column name in datain)}

\item{labelX}{a label for abscissa}
}
\value{
a graph
}
\description{
produce a graphic of the ditribution of a trait with the detected
outlier points according to a specified flag method
}
\examples{
library(dplyr)
data(plant4)
mydata<-mutate(plant4,flag=if_else(Phy <= 0.215,0,1))
outlierHist(datain=mydata,trait="Phy",
              flag="flag",labelX="Phyllocron")
}
