% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PAdata}
\alias{PAdata}
\title{Parameters of interest of plant over thermal time in an experiment}
\format{a data frame with 55799 rows and 13 variables:
\describe{
  \item{Time}{date of measurement}
  \item{plantId}{Id of the plant in the greenhouse}
  \item{Genotype}{Genotypes used in the experiment}
  \item{Treatment}{Scenario of the experiment}
  \item{Population}{panel name}
  \item{Row}{Line position in the greenhouse}
  \item{Col}{Position in the line in the greenhouse}
  \item{Biomass_Estimated}{Estimation of the biomass from the pictures}
  \item{LA_Estimated}{Estimation of the leaf area from the pictures}
  \item{Height_Estimated}{Estimation of the height from the pictures}
  \item{count_leaf}{Number of visible leaves}
  \item{phyllocron}{The phyllocron is calculated as the slope of the linear regression 
  between the number of leaves and the thermal time at 2017-04-27 day, before the
  beginning of the water deficit}
  \item{TT}{thermal time}
}}
\usage{
PAdata
}
\description{
A dataset containing biomass, plant height and leaf area over thermal time
for each plant of an experiment
}
\keyword{datasets}
