% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphicalFunctions.R
\name{plotDetectOutlierPlantMaize}
\alias{plotDetectOutlierPlantMaize}
\title{plotDetectOutlierPlantMaize}
\usage{
plotDetectOutlierPlantMaize(datain, outmodels, x, y, genotype, idColor, idFill)
}
\arguments{
\item{datain}{input dataframe, a spatio-temporal data.frame}

\item{outmodels}{the output object from funcDetectOutlierPlantMaize() function
containing the detected outliers (either small or big)
'theobject$smallOutliers' or 'theobject$bigOutliers'}

\item{x}{character, name of the time variable in datain}

\item{y}{character, name of the variable in datain to draw}

\item{genotype}{character, name of the genotype variable in datain}

\item{idColor}{character, name of the treatment variable in datain to differentiate
the curves}

\item{idFill}{character, name of the repetition variable in the datain to differentiate
the outlier curves}
}
\value{
a graphic
}
\description{
function producing a graphic of detected outlier plants from 
             funcDetectOutlierPlantMaize() function
}
\details{
see funcDetectOutlierPlantMaize function
}
\examples{
\donttest{
#plotDetectOutlierPlantMaize(datain=PAdata,
#          outmodels=test$smallOutlier,
#          x="Time",
#         y="Biomass_Estimated",
#          genotype="Genotype",
#          idColor="Treatment",
#         idFill="plantId")
}
}
