% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothingSplinesModel.R
\name{printGSS}
\alias{printGSS}
\title{a function for gss analysis description}
\usage{
printGSS(object, threshold)
}
\arguments{
\item{object}{a dataframe to describe from fitGSS() function}

\item{threshold}{numeric, a threshold for Kullback-Leibler projection}
}
\value{
a description
}
\description{
a function for gss analysis description
}
\details{
the input object is a fitGSS result and printGSS() prints the 2nd element of it (a dataframe 
         with the Kullback-Leibler projection).
         The colnames of this dataframe are Genosce, ratio, kl, check. 
         The outlier curves are identified with the Kullback-Lleiber 
         distance higher than a given threshold, see (Gu, 2014). 
         Final identification of outlier is done by an operator over
         genotypes when the test is significant.
}
\examples{
data(plant1)
selec<-c("Lo1199_H","Lo1124_H","Lo1038_H","A3_H")
mydata<-plant1[plant1[,"genotypeAlias"] \%in\% selec,]
fm1<-fitGSS(datain=mydata,trait="biovolume",loopId="genotypeAlias")
 printGSS(object=fm1,threshold=0.05)
}
\seealso{
\code{\link[=ssanova]{ssanova}}
}
