% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thermalTime.R
\name{thermalTime}
\alias{thermalTime}
\title{a function to calculate thermal time}
\usage{
thermalTime(datain, inSpecie, method, inDateS = NULL, inDateE = NULL,
  inTemp = NULL)
}
\arguments{
\item{datain}{input dataframe of meteo data from phis web service}

\item{inSpecie}{character, studied specie}

\item{method}{character, a method of thermal time's calculation ("parent","baseline")}

\item{inDateS}{a date of sowing or thinning etc... start event ("YYYY-MM-DD")}

\item{inDateE}{a date of harvesting ("YYYY-MM-DD")}

\item{inTemp}{numeric, a baseline temperature for baseline's method}
}
\value{
a dataframe
}
\description{
this function calculates the thermal time for an experiment
             according to several methods
}
\details{
Parent et. al. (2010). Modelling temperature-compensated physiological rates,
   based on the co-ordination of responses to temperature of developmental processes.
   Journal of Experimental Botany. 61 (8):2057-2069

   if the Parent's model is chosen inTemp must be NULL

   The input dataframe is extracted from phis web service (getEnvironment function)
   and is structured as follow: date, value, sensor, codeVariable and facility
}
\examples{
# Example for the model of Parent 2010
data(meteoDT)
head(meteoDT)
test<-thermalTime(datain=meteoDT,inSpecie="maize",method="parent",inDateS="2017-04-02",
                  inDateE="2017-06-15",inTemp=NULL)
head(test)
}
